+++ src/Libraries/Commands/commands_client.h
@@ -26,9 +26,9 @@ class CommandsClient final
   dequeTrim(const std::string& aContainerName, int64_t aStart, int64_t aStop);
 
   std::list<std::string> listPopFront(const std::string& aContainerName,
-                                      uint32_t           aCount);
+                                      uint64_t           aCount);
   std::list<std::string> dequePopFront(const std::string& aContainerName,
-                                       uint32_t           aCount);
+                                       uint64_t           aCount);
 
   size_t listLength(const std::string& aContainerName);
   size_t dequeLength(const std::string& aContainerName);
@@ -49,7 +49,7 @@ class CommandsClient final
 
   std::list<std::string>
   popFront(const std::string&                                   aContainerName,
-           uint32_t                                             aCount,
+           uint64_t                                             aCount,
            const std::function<::grpc::Status(::grpc::ClientContext*,
                                               const PopFrontRequest&,
                                               PopFrontReply*)>& aFunction);
